pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-businessmap"
version = "0.1.0"
description = "MCP Server for Businessmap (Kanbanize)"
readme = "README.md"
authors = [
    {name = "Seu Nome", email = "seu.email@exemplo.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "modelcontextprotocol>=0.0.8",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.237",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
mcp-businessmap = "mcp_businessmap:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
select = ["E", "F", "W", "I", "B", "C90"]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true 